Reorganizing 2022 ACC NBT ECRs v1
Deleting weighted-average ECRs (the final approved methodology uses a simple average), moves 2022 ACC NBT ECRs into their own folder. Created new folder for 2024 ACC NBT ECRs to be populated in a future commit.

This change in file location will temporarily fix the app, but this will be updated with a filepath change in the next commit.