User Inputs Update
Modified name of year input from "ACC Year" to "Export Compensation Rate Year" to clarify that this does not refer to the vintage of the ACC (which will be automatically mapped to Final Interconnection Application Year values in an upcoming commit). Changed the IX Application Year list to only cover 2023-2026 as opposed to 2023-2030. ACC Plus Adder values are defined out to 2028 and beyond, but the most recent ACC only applies to customers interconnecting in 2025-2026 so rates for customers submitting IX applications in 2027 and beyond are undefined.