2024 ACC ECRs v2
Modified both the static and interactive visualizations to use 2024 ACC ECRs if the final interconnection application year is 2025 or 2026, and the 2022 ACC ECRs if the final IX app year is 2023 or 2024. Made the default IX app year and Export Comepensation Rate Year 2025, and switched the retail-rate comparison year from 2023 to 2025.