Delivery Generation Split v1
Initial test of a new retail rate library format intended to replace the current approach to creating retail rates, with delivery and generation rates split out separately. Instead of having a separate script for each rate that converts hard-coded values into a time-series, info about all rates will be contained in a single CSV with the data in a compact "TOU Bands" format. This data will be accessed directly by the data visualization script and web app and converted into a plot-ready format.